Introduce data_structures_entry to get rid of circular imports
